backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
firebase-admin==6.2.0
gunicorn==21.2.0
python-dotenv==1.0.0
gitpython==3.1.40
markdown==3.5.1
pdfkit==1.0.0
fpdf2==2.7.6
google-generativeai==0.3.2
requests==2.31.0
werkzeug==3.0.1
